requirements.txt
discord.py==2.1.1
boto3==1.26.73
pytz==2022.7.1